requirements.txt
Faker==0.7.12
psycopg2==2.7.1
pytest==2.9.2
pytest-allure-adaptor==1.7.8
pytest-xdist==1.15.0
PyYAML==3.13
requests==2.17.3
retrying==1.3.3
urllib3==1.21.1
